requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.0
redis==5.0.0